Search for asset directory
Instead of assuming that the assets are located in the current working
directory, automatically find the location of the assets. The idea is to
start at the directory containing the executable and go upwards until
one of the assets are found ("Snippy-1.png"). Then we assume the rest of
the assets are located in the same place.

Fixes #1