Automatically determine the asset directory
We currently expect the executable to be run in the directory containing
all of the assets. Now, the program automatically determines where the
assets are located. The approach is to begin at the directory containing
the executable and move up until we find a "test asset". Then we assume
that the rest of the assets are located under the same directory.