Rewrite the `Platform` class
There's no reason to use an SFML `ConvexShape` when we can't even perform
general convex-convex shape collisions. Use the simpler `RectangleShape`
class, and explicitly support only rectangle-rectangle collisions (to be
specific, non-rotated rectangles/AABBs).

Note that `Platform::collide` not only indicates *whether* a collision is
occurring, but also the direction in which the collision is occurring.
This is needed to properly resolve the collision.